// src/types/poi.ts
export interface POI {
  id: number;
  code: string | null;
  country: string | null;
  country_en: string | null;
  country_ja: string | null;
  state: string | null;
  state_en: string | null;
  state_ja: string | null;
  city: string | null;
  city_en: string | null;
  city_ja: string | null;
  latitude: number | null;
  longitude: number | null;
}